Fix mirror dimensions
when constrainDimensions is active and not using fixed item width

When clicking an item, the mirror is added to the container.
Calculations are then made to figure out what size the mirror should be.
However, since the mirror is already in the container, the container now
has an extra item, thus reducing the width of all of the container's items.
Hence, the width calculated is too small and doesn't reflect the true
source size.

Instead, let's initially hide the mirror, figure out the intended width,
then show the mirror.