Allow specifying an array of classNames
In order to prevent breaking changes, this adds a new method
`getClassNamesFor` while keeping the existing behavior of
`getClassNameFor` (the former returns an array, and the latter returns
the first item of the array).

There's one caveat when using this feature though. Since
`classList.contains` only accepts a single class, it's somewhat
important that the first class of each DraggableClassNames is unique.